% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCD.R
\docType{data}
\name{CCD}
\alias{CCD}
\title{Credit Card Defaults}
\format{A dataset with 29728 rows and 25 variables. The variables are
as follows:
\describe{
  \item{default.payment.next.month}{Factor indicating default}
  \item{LIMIT}{Credit card limit}
  \item{SEX}{Factor indicating gender}
  \item{Education}{Factor depicting highest level of education}
  \item{Marriage}{Factor indicating marital status}
  \item{AGE}{Age}
  \item{REPAID.SEPT}{Factor describing September payment status}
  \item{REPAID.AUG}{Factor describing August payment status}
  \item{REPAID.JULY}{Factor describing July payment status}
  \item{REPAID.JUNE}{Factor describing June payment status}
  \item{REPAID.MAY}{Factor describing May payment status}
  \item{REPAID.APR}{Factor describing April payment status}
  \item{BILL_SEPT}{Total credit card bill in September}
  \item{BILL_AUG}{Total credit card bill in August}
  \item{BILL_JULY}{Total credit card bill in July}
  \item{BILL_JUNE}{Total credit card bill in June}
  \item{BILL_MAY}{Total credit card bill in May}
  \item{BILL_APR}{Total credit card bill in April}
  \item{PAYMENT_SEPT}{Total paid in September}
  \item{PAYMENT_AUG}{Total paid in August}
  \item{PAYMENT_JULY}{Total paid in July}
  \item{PAYMENT_JUNE}{Total paid in June}
  \item{PAYMENT_MAY}{Total paid in May}
  \item{PAYMENT_APR}{Total paid in April}
  }
  @source Lichman, M. (2013). UCI Machine Learning Repository [http://archive.ics.uci.edu/ml]. Irvine, CA: University of California, School of Information and Computer Science.}
\usage{
CCD
}
\description{
A dataset containing information on default payments, demographic infomration,
and payment history for Taiwanese credit card clients from April 2005 to 
Septembrer 2005
}
\details{
This dataset was retrieved by UCI's Machine Learning Repository and
then slightly modified to improve the predictiveness of the variables.
}
\keyword{datasets}
